+++ CHANGELOG.md
@@ -4,6 +4,10 @@ Format based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/)
 
 ## [Unreleased]
 
+### Added
+
+* `--sources` to show definitions' source dictionary
+
 ### Changed
 
 * Updated dependencies